fix:rework ewm-service
fix:rework ewm-service

fix:rework ewm-service

fix:rework ewm-service

+++ ewm-service/src/main/java/ru/practicum/model/event/NewEventDto.java
@@ -23,7 +23,7 @@ public class NewEventDto {
     @NotNull
     Location location;
     Boolean paid;
-    Integer participantLimit;
+    Long participantLimit;
     Boolean requestModeration;
     @NotNull
     String title;
